+++ src/rules/no-duplicate-selectors/index.js
@@ -1,5 +1,8 @@
-import { includes } from "lodash"
+import { includes, union } from "lodash"
+import resolvedNestedSelector from "postcss-resolve-nested-selector"
 import {
+  cssNodeContextLookup,
+  findMediaContext,
   ruleMessages,
   report,
   validateOptions,
@@ -15,7 +18,7 @@ export const messages = ruleMessages(ruleName, {
 // Each source maps to another map, which maps rule parents to a set of selectors.
 // This ensures that selectors are only checked against selectors
 // from other rules that share the same parent and the same source.
-var selectorContextMap = new Map()
+var selectorContextLookup = cssNodeContextLookup()
 
 export default function (actual) {
   return (root, result) => {
@@ -23,36 +26,34 @@ export default function (actual) {
     if (!validOptions) { return }
 
     root.walkRules(rule => {
-      const ruleSource = rule.source.input.from
-      const ruleParent = rule.parent
-      if (!selectorContextMap.has(ruleSource)) {
-        selectorContextMap.set(ruleSource, new Map())
-      }
-      if (!selectorContextMap.get(ruleSource).has(ruleParent)) {
-        selectorContextMap.get(ruleSource).set(ruleParent, new Set())
-      }
-      const parentSelectorSet = selectorContextMap.get(ruleSource).get(ruleParent)
-
-      const normalizedSelectorList = rule.selectors.map(normalizeSelector)
+      const contextSelectorSet = selectorContextLookup.getContext(rule, findMediaContext(rule))
+      const resolvedSelectors = rule.selectors.reduce((result, selector) => {
+        return union(result, resolvedNestedSelector(selector, rule))
+      }, [])
+      const normalizedSelectorList = resolvedSelectors.map(normalizeSelector)
 
       // Complain if the same selector list occurs twice
 
       // Sort the selectors list so that the order of the constituents
       // doesn't matter
       const sortedSelectorList = normalizedSelectorList.slice().sort().join(",")
-      if (parentSelectorSet.has(sortedSelectorList)) {
+      if (contextSelectorSet.has(sortedSelectorList)) {
+        // If the selector isn't nested we can use its raw value; otherwise,
+        // we have to approximate something for the message -- which is close enough
+        const isNestedSelector = resolvedSelectors.join(",") !== rule.selectors.join(",")
+        const selectorForMessage = (isNestedSelector) ? resolvedSelectors.join(", ") : rule.selector
         return report({
           result,
           ruleName,
           node: rule,
-          message: messages.rejected(rule.selector),
+          message: messages.rejected(selectorForMessage),
         })
       }
 
-      parentSelectorSet.add(sortedSelectorList)
+      // We're treating the Map created by cssNodeContextLookup as a Set
+      contextSelectorSet.set(sortedSelectorList, null)
 
       // Or complain if one selector list contains the same selector more than one
-
       rule.selectors.forEach((selector, i) => {
         if (includes(normalizedSelectorList.slice(0, i), normalizeSelector(selector))) {
           report({